// demos/expense-approval-automation/src/generated/services/Expaa_expensesService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Expaa_expensesBase, Expaa_expenses } from '../models/Expaa_expensesModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Expaa_expensesService {
  private static readonly dataSourceName = 'expaa_expenses';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Expaa_expensesBase, 'expaa_expenseid'>): Promise<IOperationResult<Expaa_expenses>> {
    const result = await Expaa_expensesService.client.createRecordAsync<Omit<Expaa_expensesBase, 'expaa_expenseid'>, Expaa_expenses>(
      Expaa_expensesService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Expaa_expensesBase, 'expaa_expenseid'>>): Promise<IOperationResult<Expaa_expenses>> {
    const result = await Expaa_expensesService.client.updateRecordAsync<Partial<Omit<Expaa_expensesBase, 'expaa_expenseid'>>, Expaa_expenses>(
      Expaa_expensesService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Expaa_expensesService.client.deleteRecordAsync(
      Expaa_expensesService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Expaa_expenses>> {
    const result = await Expaa_expensesService.client.retrieveRecordAsync<Expaa_expenses>(
      Expaa_expensesService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Expaa_expenses[]>> {
    const result = await Expaa_expensesService.client.retrieveMultipleRecordsAsync<Expaa_expenses>(
      Expaa_expensesService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Expaa_expenses> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Expaa_expensesService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Expaa_expensesService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}