// demos/expense-approval-automation/src/generated/services/Expaa_approvalsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Expaa_approvalsBase, Expaa_approvals } from '../models/Expaa_approvalsModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Expaa_approvalsService {
  private static readonly dataSourceName = 'expaa_approvals';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Expaa_approvalsBase, 'expaa_approvalid'>): Promise<IOperationResult<Expaa_approvals>> {
    const result = await Expaa_approvalsService.client.createRecordAsync<Omit<Expaa_approvalsBase, 'expaa_approvalid'>, Expaa_approvals>(
      Expaa_approvalsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Expaa_approvalsBase, 'expaa_approvalid'>>): Promise<IOperationResult<Expaa_approvals>> {
    const result = await Expaa_approvalsService.client.updateRecordAsync<Partial<Omit<Expaa_approvalsBase, 'expaa_approvalid'>>, Expaa_approvals>(
      Expaa_approvalsService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Expaa_approvalsService.client.deleteRecordAsync(
      Expaa_approvalsService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Expaa_approvals>> {
    const result = await Expaa_approvalsService.client.retrieveRecordAsync<Expaa_approvals>(
      Expaa_approvalsService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Expaa_approvals[]>> {
    const result = await Expaa_approvalsService.client.retrieveMultipleRecordsAsync<Expaa_approvals>(
      Expaa_approvalsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Expaa_approvals> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Expaa_approvalsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Expaa_approvalsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}