// demos/expense-approval-automation/src/generated/models/Expaa_expensesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Expaa_expensesexpaa_category = {
  1: 'Travel',
  2: 'Meals',
  3: 'Lodging',
  4: 'Supplies',
  5: 'Software',
  6: 'Training',
  7: 'Other'
} as const;
export type Expaa_expensesexpaa_category = keyof typeof Expaa_expensesexpaa_category;
export const Expaa_expensesexpaa_paymentmethod = {
  1: 'PersonalCard',
  2: 'CorporateCard',
  3: 'Cash',
  4: 'BankTransfer'
} as const;
export type Expaa_expensesexpaa_paymentmethod = keyof typeof Expaa_expensesexpaa_paymentmethod;
export const Expaa_expensesexpaa_status = {
  1: 'Submitted',
  2: 'InReview',
  3: 'Approved',
  4: 'Rejected',
  5: 'Reimbursed'
} as const;
export type Expaa_expensesexpaa_status = keyof typeof Expaa_expensesexpaa_status;
export const Expaa_expensesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Expaa_expensesstatecode = keyof typeof Expaa_expensesstatecode;
export const Expaa_expensesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Expaa_expensesstatuscode = keyof typeof Expaa_expensesstatuscode;

export interface Expaa_expensesBase {
  expaa_amount?: number;
  expaa_approvername?: string;
  expaa_category?: Expaa_expensesexpaa_category;
  expaa_decisionnotes?: string;
  expaa_decisionon?: string;
  expaa_description?: string;
  expaa_employeeemail?: string;
  expaa_employeename?: string;
  expaa_expensedate?: string;
  expaa_expenseid: string;
  expaa_merchant?: string;
  expaa_name?: string;
  expaa_paymentmethod?: Expaa_expensesexpaa_paymentmethod;
  expaa_policyflag?: boolean;
  expaa_policynotes?: string;
  expaa_receipturl?: string;
  expaa_status?: Expaa_expensesexpaa_status;
  expaa_submittedon?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Expaa_expensesstatecode;
  statuscode?: Expaa_expensesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Expaa_expenses extends Expaa_expensesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  expaa_categoryname?: string;
  expaa_paymentmethodname?: string;
  expaa_statusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}